move feature jsx into components
Refs #197
