+++ docs/architecture.md
@@ -44,12 +44,14 @@ These paths are implemented as Next.js Route Handlers under `src/app/api/proxy/*
 
 API routes under `src/app/api/**/route.js` are HTTP adapters. They parse request parameters, enforce proxy/security policy, call a domain mechanism, and translate the mechanism result into `Response` or `NextResponse`.
 
-Domain code lives under `src/features/<domain>/`, including browser-facing feature UI and server-facing API mechanisms:
+Functionality-level domain code lives under `src/features/<domain>/` as plain `.js` modules:
 
 - `<domain>.mechanism.js` owns source selection, fallback order, cache policy, request parameterization, and provider orchestration.
 - `<domain>.models.js` owns domain constants, result metadata, and mechanism-specific error types.
 - `<domain>.utils.js` owns pure normalization and predicate helpers.
 - Prefix families are grouped by product concept, e.g. `src/features/aircraft/*`, `src/features/airport/*`, `src/features/aviation/flight-routes`, and `src/features/weather/metar`.
+
+JSX components live under `src/components/**`, grouped by screen or product domain. Components may import feature modules, but feature modules should not import JSX components.
 
 Persistence boundaries live under `src/app/api/dao/*.dao.js`. DAO files should contain Supabase/SQL reads and writes only; they should not choose providers, cache policy, fallback behavior, or import mechanism files.
