+++ src/components/aircraft/preview/AircraftPreviewMediaCard.jsx
