+++ src/components/aircraft/preview/AircraftPreviewMetadataCard.jsx
