+++ src/components/aircraft/preview/AircraftPreviewType.jsx
