+++ src/components/map/AirportMap.jsx
@@ -13,8 +13,8 @@ import SelectedAircraftTrace from "./SelectedAircraftTrace.jsx";
 import RunwayAnnotationLayer from "./RunwayAnnotationLayer.jsx";
 import ProcedureSegmentLayer from "./ProcedureSegmentLayer.jsx";
 import { AIRPORT_MAP_FALLBACK_CENTER } from "../../config/airportMap.js";
-import MapAttribution from "../../features/airport/map/MapAttribution.jsx";
-import MapLoadingState from "../../features/airport/map/MapLoadingState.jsx";
+import MapAttribution from "./MapAttribution.jsx";
+import MapLoadingState from "./MapLoadingState.jsx";
 import { getAircraftIdentity } from "../../features/airport/context/airportContextUiModel.js";
 import { aircraftMatchesFilters } from "../../features/aircraft/filters/aircraftFilters.js";
 import {
