+++ src/components/map/MapCoordinateLabel.jsx
