+++ src/components/map/controls/MapLayerDrawer.jsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Button } from "../../../components/ui/button.jsx";
+import { Button } from "@/components/ui/button.jsx";
 import { MapControlIcon } from "./mapControlIcons.jsx";
 
 const LAYER_TOGGLES = [
