+++ src/components/map/controls/mapControlIcons.jsx
