+++ src/components/screens/AboutClient.jsx
@@ -1,5 +1,5 @@
 "use client";
 
-import AboutPanel from "../../features/about/AboutPanel.jsx";
+import AboutPanel from "@/components/about/AboutPanel.jsx";
 
 export default AboutPanel;
